Remove loading of overdue item count on the home screen
This is being done because the overdue query is quite complex and we
load both the count of items and a paginated list of the items when
we open the open screen which causes a slowdown in future, immediately
required database queries.

This commit disables showing the overdue count temporarily until we
figure out how to improve the app performance.