Make `PatientTeleconsultationInfo` to be a sealed class
We would want different resources for different message builders, in this case for short message builder we only need patient id and teleconsult id so there is no point loading rest of the resources. So instead of making rest of the resources nullable, we are using sealed calsses to have different `PatientTeleconsultationInfo`s. One downside for this approach is message builders will take any `PatientTeleconsultationInfo`s and since they are force casted to specific types in their message builders, that can cause crash if developer passed a wrong `PatientTeleconsultationInfo` type.