Check the current screen before updating the app lock time (#1871)
* Add check for current screen before updating the app lock timestamp

* Change `ConfirmResetPinDialog` to push the `ForgotPinCreateNewPinScreen` onto the current stack

We currently check if the `AppLockScreen` is in the current stack before
deciding whether to update the app lock time or not. If we clear the
stack as we were doing earlier, then it is possible to bypass the app
lock screen by opening the confirm pin reset screen, exiting the app, and
then launching it again.

There are two ways we can solve this:
1. Maintain a list of all screens for which the app lock time should *not* be updated
2. Maintain the current stack with the `AppLockScreen` instead of clearing it

I decided to go for approach #2 since I felt it required less maintenance
in the long run since approach #1 would require us to update the list of
screens if it changes.

* Update CHANGELOG

(cherry picked from commit dc36d755ed6382a45e4890584f3705ee86824895)