Skip triggering a purge if there is no user present locally
Currently, the app triggers a complete sync even if the user is not
logged in during the login flow. Since the purging of unused data
requires access to the user's current facility and there is no local
user present, this causes a crash since the current facility cannot be
fetched.

A long term fix is to move the login flow to a separate activity and
trigger the syncs only in the activity where we can be guaranteed that
a user will be present locally. However, that will require considerable
effort, so for the short term, we change it so that the sync checks for
a local user being present before triggering a purge.