Bottom button bar layout is broken on the patient summary screen on some devices (#2040)
* Fix teleconsult button alignment in patient summary screen

Since we were using linear layout with padding initially, we had to set `clipToPadding=false` in order for the shadow on button press to be not get clipped. Because of that the button get's misaligned when the text inside the button changes to multiline. In order to fix this issue and align the button to center, I have switched to using ConstraintLayout.

* Update CHANGELOG