Set visibility of `prescribeddrugs_done` button to gone
Currently the save button is shown initially and blinks to switch to refill medicines, in order to avoid that I have set both the buttons visibility to gone and ui render will decide which buttons to show. Also, I have removed padding because it causes the button frame to be shown when both buttons are not visible, we don't want it to happen.