Fix BP passport scanning not working (#2167)
* Set target aspect ration for the image analyzer

In some scenarios it might be possible certain devices send a large image to image analyser which will cause the analyser to fail. In order to resolve this we calculate the aspect ratio from the view finder/preview dimensions and set the target aspect ratio. Just to be safe we are also setting the target orientation to avoid any orientation related scanning issues.

* Update CHANGELOG