Remove explicit back press handling in `AccessDeniedScreen`
This is being done because the access denied screen is being used in
two different activities (`AuthenticationActivity` and `TheActivity`).

Since we are moving the activities to the new navigation framework in
pieces, we will not be able to have two different methods of back
press handling in the same screen since the `ScreenRouter` and the
`Router` cannot be injected into the class simultaneously.

Since we already clear the history whenever we push the access denied
screen into the stack in both the activities, this additional back press
handling is redundant and can be removed.