Disable smooth scroll to position in instant search
It's possible the user might have scrolled  really long in a list and when they update the search query the recycelrview will smooth scroll which can taka a while depending on how many results they have scrolled past. To avoid that we are scrolling fast to first item without any animation.