Fetch the initial screen in `TheActivity#onCreate` instead of in `onAttachBaseContext`
Due to the differences in how the `Router` and `ScreenRouter` setup
the initial screen, the `UserSession` instance is no longer available
at the time of deciding the initial screen.

This commit does a workaround by loading an `EmptyScreenKey` as the
initial screen and then loads the initial screen in `onCreate`. This
is a temporary solution and the right way to solve this is to move the
loading of the initial screen to the  Mobius loop, which is being
tracked here:

https://app.clubhouse.io/simpledotorg/story/2342/move-loading-of-the-initial-screen-in-theactivity-to-the-mobius-loop