Trigger the effect to go back only once in `EnterOtpScreen`
This was being done because it was possible for the effect to be emitted
twice. Once when the user enters the login OTP, and again when the user
gets verified in the background.

With the new navigation framework, this seems to create a situation where
the framework executes the `.pop()` method twice and causes a crash
because the framework does not allow a `.pop()` when there's only a
single screen in the stack.