Material Theming Migration - 4 (#2300)
* Add toolbar button style

* Use new Simple toolbar button style

* Remove unused resources related to toolbar button

* Change `colorSecondaryVariant` color

We are not using `simple_green_200` anywhere, so we removed that and started using `simple_green_100` as `colorSecondaryVariant`. That way we will have consisten styling for button frame for color primary button and color secondary buttons.

* Use `color_on_toolbar_primary_72` as text color for `bangladeshNationalIdTextView`

* Rename `white100` usage to `white`

* Replace yellow2 usage with `simple_yellow_100`

* Rename drawable names used in `contactpatient_callpatient`

* Clean up user status snackbars

* Use chips for medical history answers

* Add input type to pin entry styles

* Fix error state visibility in `screen_registratio_confirm_pin`

* Fix `introVideoSubtitle` text color for devices below API 23

* Use separate button frame for log teleconsult button

Instead of setting the frame color at runtime. We will be using a separate button frame with a theme overlay to change the colors

* Use theme attribute for setting full screen key changer background

* Move `facility_search_query_highlight` to use new Simple colors

* Set `identifierTextView` drawable tint in layout XML

* Clean up `InstantSearchResultsItemType`

* Fix search results empty arrow stem

* Clean up `ShortCodeSearchResultScreen`

* Set `android:textColor` to `null` in text appearances

Text appearances by default sets the text color to `textColorPrimary`(`colorOnPrimary`) causing our `TextAppearanceSpan` to get that color whenever we are applying a new text style. Ideally it should just maintain the text color set at the view level. Other option would be build a custom text appearance span, but that would be significantly more time consuming and complex.

* Add `TextAppearance.Simple.Caption`

* Remove color span from `PatientSummaryScreen`

The color is same as the text view. Previously we did this to avoid setting the color from text style

* Use `buildSpannedString` instead of `Truss` in `OnboardingScreen`

* Use `buildSpannedString` instead of `Truss` in `BloodPressureHistoryListItem`

* Use `buildSpannedString` instead of `Truss` in `BloodSugarHistoryListItem`

* Use `buildSpannedString` instead of `Truss` in `BloodSugarItemView`

* Use `buildSpannedString` instead of `Truss` in `BloodPressureItemView`

* Use `buildSpannedString` instead of `Truss` in `PatientEntryScreen`

* Remove `Truss`

* Apply shape appearance overlay for base bottom sheet

* Clean up `link_id_with_patient_view`

* Use new bottom sheet dialog theme in `BaseBottomSheet`

It's not necessary since we applied the default style in theme. But just to be safe I have added it here as well, once we can verify the bottom sheet with new navigation framework. We can remove this.

* Use new Simple colors instead of older ones

* Change `ThemeOverlay.Simple.Dialog` material button style

* Remove unused resources

* Update lint excludes

* Fix patients empty state background

Also center align the text

* Replace `CardView` with `MaterialCardView`

* Rename `WhitePrimary` theme overlay to `SurfacePrimary`

* Rename `colors_v3.xml` to `colors.xml`

* Rename `styles_v3.xml` to `styles.xml`

* Rename `theme_v2.xml` to `theme.xml`

* Update CHANGELOG

* Clean up `bloodsugarentry_flipper_blood_sugar_entry`

* Remove text size dimens usage

* Rename `textsize_24` to `fast_scroll_pop_up_text_size`

We can drop this once older all patients view is removed

* Update lint rules