Migrate `TheActivity` to the new navigation framework (#2294)
* Add method to `Router` to check if a key of a specific type is present

* Use `Router` in `TheAcivity` to manage the screen backstack

* Add `ScreenResultBus` as a factory dependency to `TheActivity`

* Remove `ScreenRouter` from `MedicalHistorySummaryView`

* Use `Router` in `DrugSummaryView`

* Use `Router` in `BloodSugarSummaryView`

* Use `Router` in `BloodPressureHistoryScreen`

* Use `Router` in `BloodPressureSummaryView`

* Use `Router` in `BloodSugarHistoryScreen`

* Use `Router` in `EnterOtpScreen`

* Use `Router` in `DeletePatientScreen`

* Use `Router` in `PatientsTabScreen`

* Use `Router` in `HelpScreen`

* Use `Router` in `RecentPatientsView`

* Use `Router` in `AssignedFacilityView`

* Use `Router` in `RecentPatientsScreen`

* Use `Router` in `ForgotPinCreateNewPinScreen`

* Use `Router` in `OverdueScreen`

* Use `Router` in `PatientSearchResultsScreen`

* Use `Router` in `PatientSearchScreen`

* Use `Router` in `ForgotPinConfirmPinScreen`

* Use `Router` in `HomeScreen`

* Use `Router` in `PatientEntryScreen`

* Use `Router` in `EditPatientScreen`

* Use `Router` in `AppLockScreen`

* Remove `ScreenRouter` from `PatientSearchView`

* Use `Router` in `TeleconsultSuccessScreen`

* Use `Router` in `TeleconsultRecordScreen`

* Use `Router` in `TeleconsultRecordedDialog`

* Use `Router` in `TeleconsultPrescriptionScreen`

* Remove `ScreenRouter` from `TeleconsultPatientInfoView`

* Use `Router` in `TeleconsultMedicinesView`

* Remove `ScreenRouter` from `TeleconsultDoctorInfoView`

* Use `Router` in `PatientSummaryScreen`

* Use `Router` in `ConfirmResetPinDialog`

* Use `Router` in `NewMedicalHistoryScreen`

* Use `Router` in `ConfirmDiscardChangesDialog`

* Use `Router` in `EditMedicinesScreen`

* Use `Router` in `ShortCodeSearchResultScreen`

* Use `Router` in `SettingsScreen`

* Use `Router` in `ChangeLanguageScreen`

* Use `Router` in `TeleconsultSharePrescriptionScreen`

* Use `Router` in `InstantSearchScreen`

* Fetch the initial screen in `TheActivity#onCreate` instead of in
`onAttachBaseContext`

Due to the differences in how the `Router` and `ScreenRouter` setup
the initial screen, the `UserSession` instance is no longer available
at the time of deciding the initial screen.

This commit does a workaround by loading an `EmptyScreenKey` as the
initial screen and then loads the initial screen in `onCreate`. This
is a temporary solution and the right way to solve this is to move the
loading of the initial screen to the  Mobius loop, which is being
tracked here:

https://app.clubhouse.io/simpledotorg/story/2342/move-loading-of-the-initial-screen-in-theactivity-to-the-mobius-loop

* Change `InstantSearchScreen` to a fragment

* Change `ScanSimpleIdScreen` to a fragment screen

* Change `HomeScreen` to a fragment screen

This commit also changes the scanning of BP Passports to be via the
new navigation flow.

* Remove `ScanBpPassportActivity`

This was done because the older method of passing the scanning results
would not work since this activity was a full screen activity and would
cause the previous activity to be restarted and drop the activity result
event.

* Remove deprecated Mobius delegate constructor in `BloodPressureSummaryView`

This was done because the older reference to `delegate.prepare()` was
causing the delegate to be instantiated in `onFinishInflate()` which
led to a crash because the fragment was not yet associated with it.

* Remove deprecated Mobius delegate constructor in `BloodSugarSummaryView`

This was done because the older reference to `delegate.prepare()` was
causing the delegate to be instantiated in `onFinishInflate()` which
led to a crash because the fragment was not yet associated with it.

* Trigger the effect to go back only once in `EnterOtpScreen`

This was being done because it was possible for the effect to be emitted
twice. Once when the user enters the login OTP, and again when the user
gets verified in the background.

With the new navigation framework, this seems to create a situation where
the framework executes the `.pop()` method twice and causes a crash
because the framework does not allow a `.pop()` when there's only a
single screen in the stack.

* Update CHANGELOG

* Fix failing unit tests for finding the initial screen key