Send BP Passport sheet results via the additional events mechanism
This was being done because the Mobius loop in `InstantSearchScreen`
was being started after the activity result from `BpPassportSheet` was
delivered, which resulted in the event never getting emitted.

Moving this to the additional events mechanism is a temporary workaround
until we can migrate `BpPassportSheet` fully to the new navigation.