Delay changing the navigation stack by a little bit so the current fragment transaction can complete
We currently need to close the sheet and proceed to the next screen if
the current facility was not switched. This was being done in the `onViewCreated()`
method. However, with the new navigation framework, this gets invoked
while the current fragment transaction is ongoing, which causes a crash.

The right way to fix this is to move this check into the Mobius loop
and get rid of the delayed job, but this workaround should suffice for
now.