Don't assign a reviewer if a reviewer already exist
In certain scenarios it's possible the auto request review action will assign reviewer even if one already is present. For example, when we mark a PR as draft without removing the reviewer and mark it ready for review, the auto request review action will run and tries to add another reviewer.