Remove automatic scrolling to top when setting/updating adapter data
We want to preserve the scroll position in search when user opens patients summary and comes back to search, so we should not be manually scrolling to top everytime we are setting or updating the data. When the query changes the data gets reordered already, so we don't have to manually reset it.