Paginate search results in `InstantSearchScreen` (#2706)
* Add query to load search by name results as paging source

* Add query to load search by number results as paging source

* Load search results as paging source

* Show patient search results using `PagingItemAdapter`

* When search results are loaded, then show search results

* Add load state listener to search results adapter

* Remove automatic scrolling to top when setting/updating adapter data

We want to preserve the scroll position in search when user opens patients summary and comes back to search, so we should not be manually scrolling to top everytime we are setting or updating the data. When the query changes the data gets reordered already, so we don't have to manually reset it.

* Remove the default adapter set in `onViewCreated`

This causes issues with scroll position restoration, so instead we will wait for the correct adapter to be set once the page is opened again.

* Show and hide progress bar when the refresh load state is loading

`Refresh` state is what get's changed when we are loading the data for the first time (this doesn't not include loading when user reaches end of pagination and pager is fetching more pages)

* Remove `InstantSearchProgressState` from `InstantSearchModel`

Since we are now using the load state listeners to show and hide progress states, we no longer have to use this model to track the progress state. We are going to probably reintroduce a different way to update the progress state in model in later commits.

* Remove effects to hide and show search no search results view

This is handled by the adapter load state listeners

* Remove `InstantSearchResultsItemType_old`

* Remove `PatientRepository#search`

* Remove search by patient identifier feature flag and make it default

* Remove `PhoneNumberCriteria`

* Update CHANGELOG