Fix `ContactPatientBottomSheet` UI spacing and styling (#2846)
* Rename `Widget.Simple.PatientContactCallResultItem` to `Widget.Simple.PatientContactCallResultItem_Old`

* Add `Widget.Simple.PatientContactCallResultItem`

* Set `colorBackground` in `ThemeOverlay.Simple.BottomSheetDialog`

The default theme overlay we are extending seems to override `colorBackground` internally which is causing issues when trying to access our parent theme `colorBackground`. I have verified if anything breaks if we make this change and so far everything is working as expected.

* Fix `contactpatient_callpatient` UI spacing and styling

* Update CHANGELOG

* Rename `contactpatient_patient_transferred_from` string to `contactpatient_patient_transfer_from`

Change the string value to `Transfer from`