Use `MobiusLoopViewModel` for sending view effects that are lifecycle aware (#2857)
* Add view effects generic type for `BaseScreen`

This will be used as type for view effect type in Mobius loop view model

* Pass `Any` as view effect type when extending `BaseScreen`

Currently we don't have a specific type for view effects, so instead of creating new type and handling view effects. For now we are going to ignore view effects for existing screens and slowly migrate them one by one.

* Add `MobiusLoopViewModel` in `BaseScreen`

* Remove Mobius controller usage from `BaseScreen`

* Forward UI events to ViewModel

* When ViewModel model updates, then render UI

* Change `BaseScreen#createEffectHandler` signature to accept view effect consumer

We can use this consumer to pass view effects into Mobius loop, similar to how we pass UI events with Mobius

* Add `NoopViewEffectHandler`

* Handle view effects in `BaseScreen`

The second observer we are adding in the `viewEffects.setObserver` is for paused views effects when app is in background. This will be invoked once lifecycle is resumed in the order the view effects are received.

https://github.com/spotify/mobius/blob/93da582db985e381f9cc11e23e026eb202e7652d/mobius-android/src/main/java/com/spotify/mobius/android/LiveQueue.java#L69

* Dispose of subscribed events in `BaseScreen`

* Update CHANGELOG

* Rename `_viewModel` to `viewModel` in `BaseScreen`