Restart `TheActivity` after changing language (#2883)
* Restart `TheActivity` instead of recreating

Recreating activity doesn't recreate the `Router` or the fragments inside it. Causing the app to crash when some operations that require context are performed like navigation, view access, etc., In order to prevent that, we need to restart the activity properly instead of just recreating it.

One downside of this will be, we take the user back to the home screen after language is selected instead of staying in language selection screen. Which I think is a good tradeoff to make to fix the crash.

* Update CHANGELOG