Use `pluginManagement` DSL for managing plugins version
We cannot directly reference the plugins in the TOML file, and we cannot use the generated `libs` in the `plugins` block. So we are going with a different way of managing plugins.

refer: https://melix.github.io/blog/#_can_i_use_a_version_catalog_to_declare_plugin_versions

Support for it is merged in and should be available in 7.2

https://github.com/gradle/gradle/pull/17394