Implement drug frequency dialog in `CustomDrugEntrySheet` (#3002)
* Create `SelectDrugFrequencyDialog`

* Refactor `EditFrequencyClicked` to a object and pass frequency value from the model

* Show edit frequency dialog when edit frequency is clicked

* Update frequency on successful dialog result

* Pass `nullIfBlank()` with dosage to avoid dosage values that are null and blank

* when frequency is edited and has unknown or null then

* Add `dosagePlaceHolder` in `CustomDrugEntryModel`

* Update CHANGELOG

* Fix `removeButton` padding and alignment

* Fix `titleTextView` margins

* Use constraint layout chains to maintain a flat view hierarchy

* Set `colorSecondary` for `ThemeOverlay.Simple.MaterialAlertDialog`

The default theme for Material alert dialog was picking up the wrong color for the radio buttons so previously we had to pass the style while building the dialog. So by setting the color secondary we solved the problem.

* Send back the result when drug frequency is selected and dismiss the dialog when done button is clicked


We have removed the mutable state where we were tracking the `selectedValueIndex` when the done button was being pressed.

* Use `AppCompatDialogFragment` instead of `DialogFragment` in `SelectDrugFrequencyDialog`

* Change `FrequencyEdited` to support null values

* Return the successful drug frequency selected result using `SelectedDrugFrequency`

Initially we were sending back `Unknown` when no drug frequency was selected which was incorrect because unknown does not mean that no frequency is selected but it can also mean that we are not supporting a newly added frequency type in previous versions.

* Use a list of `DrugFrequencyChoiceItem` to represent the radio group items in the dialog

Initially, we were using a hard coded sting-array to represent a drug array for the dialog but that was not a good approach to handle this. Hence we created a data class which will convert the `DrugFrequency` to respective strings and vice versa.

* Make the drug place holder string res value as non translatable

* Constraint `removeButton` layout to be in the right and vertical center of `titleTextView`

* Remove `SetSheetTitle` effect

* Set `SetSheetTitle` in the Ui renderer instead of calling a ui action from the effect handler

UiRenderer renders the ui every time the model is changed so setting it from the effect handler was not necessary.