Add v2 manifest support in app (#3008)
* Add `Deployment` data class

* Add `CountryV2` data class

* Load support countries v2 list from manifest

We are going to eventually replace this with the v1 `Country`. One issue would be, replacing the v1 `Country` with a v2 `Country` can cause unindented behaviour, since we will have to replace the v1 `Country` in preferences with v2 one. We need to figure out a way to migrate v1 `Country` to v2.

* Update CHANGELOG