Handle deep link results in `HomeScreen`
Technically our app starting point for user flow is `HomeScreen`, so instead of handling deep link results in `TheActivity` where it's possible for the patient summary or error dialogs to show up before app lock screen. We will handle the result in `HomeScreen` which is shown after the app lock screen.