Fix patient summary going back to the home screen when opening teleconsult links (#3161)
* Add `Intent.popWithParcelableResult` extension

* Handle deep link results in `HomeScreen`

Technically our app starting point for user flow is `HomeScreen`, so instead of handling deep link results in `TheActivity` where it's possible for the patient summary or error dialogs to show up before app lock screen. We will handle the result in `HomeScreen` which is shown after the app lock screen.

* Remove handling deep link results in `TheActivity`

* Update CHANGELOG