Fix `Country` v2 migration not running (#3183)
* Remove new country check when creating v2 country from old country

Setup activity ends up creating fallback country, so this migration ends up never running. To avoid that, we are running the migration only if the old country preference is present. Once migration is done, we delete that key, so this check should run only once.

* Remove new country check when creating deployment from the old country

Setup activity ends up creating fallback country, so this migration ends up never running. To avoid that, we are running the migration only if the old country preference is present. Once migration is done, we delete that key, so this check should run only once.

* Update CHANGELOG