Fix flaky test in `AppointmentRepositoryAndroidTest`
The test would occasionally fail because of different ordering of elements than what was expected. What was happening was that the blood sugar and pressure readings would be randomly generated in tests, which would lead to the appointment being considered as "high risk" in some cases and change the ordering from what was expected.

This commit fixes the test be ensuring the measurement readings are fixed for all the appointments.