Mock `InputFieldsFactory` in `EditPatientEffectHandlerTest`
Since we are now using a lambda for parsing dates in `DateOfBirthField`. everytime `provideFields` is called the instance of it changes and the assertion fails. One way we can handle this is to provide a list using mockito to control the return instance. We can eventually just drop the `InputFieldsFactory` all together and just provide the list of fields directly from dagger. Will do that in a different PR.