Use view render to setup UI and fill fields in edit screen (#3338)
* Change `InputField` as `sealed class`

This allow us to annotate the child classes with `Parcelize` and place the list of `InputField`s in a UI model.

* Use lambda function to parse date in `DateOfBirthField#validate`

We are doing this, so that we can mark these classes as `Parcelable`

* Make `InputField`s parcelable

* Make `InputFields` parcelable

* Add `InputFields` to `EditPatientModel`

* Mock `InputFieldsFactory` in `EditPatientEffectHandlerTest`

Since we are now using a lambda for parsing dates in `DateOfBirthField`. everytime `provideFields` is called the instance of it changes and the assertion fails. One way we can handle this is to provide a list using mockito to control the return instance. We can eventually just drop the `InputFieldsFactory` all together and just provide the list of fields directly from dagger. Will do that in a different PR.

* Add `InputFields` to `EditPatientModel`

* When input fields are loaded, then setup the ui

* Remove dispatching `SetupUi` view effect, when input fields are loaded

* Remove `SetupUi` view effect

* Add bp passports to `EditPatientModel`

* When BP passports are loaded, then update the model

* Set default visibility of `bpPassportsLabel` to gone

* When BP passports are loaded, then display identifiers

* Remove dispatching view effect to display BP passports, when BP passports are loaded

* Remove `DisplayBpPassportsEffect` view effect

* Disable state saving in text fields in `EditPatientScreen`

We don't particularly have to do this, but makes it easier to identify any potential issues and makes the view rendering more composable.

* Add `afterTextChangedWatcher` helper function to create text watcher easily

While AndroidX offers a `TextView` extension called `doAfterTextChanged` which returns a `TextWatcher`, it automatically adds that to `TextView`. This function just acts a synthetic sugar to create a `TextWatcher` instead

* Use `TextWatcher` to listen to text field changes in `EditPatientScreen`

We are doing this so that we can control adding and removing the text watcher for the text field. This would allow us to prefill the content in view render and avoid creating multiple anonymous classes or a loop of updates which can freeze the UI

* Add `EditText` extension to set text with a text watcher

* Fill form fields when model changes

* Stop prefilling form fields when edit patietn screen is created

* Remove `PrefillFormEffect` view effect

* Add `RadioGroup#check(@IdRes Int, OnCheckedChangeListener)` extension

* Use `RadioGroup#check` for toggling selected gender in `EditPatientScreen`

* Update CHANGELOG