When alternative id is loaded, then display the alternate ID
We are doing this to ensure we are setting the text watcher on the edit text field. Without it, changes will not be updated in the model.

Broken down alternative id rendering into 2 separate methods, so that it's easier to test and also make sure we are handling all identifier types properly.