Fix text change events not triggering in edit patient screen (#3343)
* Call `setColonyOrVillage` even if the entry is blank

We are doing this to ensure we are setting the text watcher on the edit text field. Without it, changes will not be updated in the model.

* Call `setPatientPhoneNumber` even if the entry is blank

We are doing this to ensure we are setting the text watcher on the edit text field. Without it, changes will not be updated in the model.

* When alternative id is loaded, then display the alternate ID

We are doing this to ensure we are setting the text watcher on the edit text field. Without it, changes will not be updated in the model.

Broken down alternative id rendering into 2 separate methods, so that it's easier to test and also make sure we are handling all identifier types properly.

* Update CHANGELOG