Show next appointment card in patient summary (#3370)
* Show next appointment card in patient summary screen

* Add query to fetch prescribed drugs count

* Handle `LoadPatientRegistrationData` effect

* Track `hasPatientRegistrationData` in `PatientSummaryModel`

We are only tracking the boolean value derived from the counts, since we don't need the individual count for any other state, so I don't see the reason to track it in model right now.

It might change in future, if we need the individual count.

* Render patient summary screen when `hasPatientRegistrationData` is not null

We are doing this to ensure the data is loaded and ready to render

* When patient registration data is loaded, then update the model

* Show next appointment card, when patient registration data is present

* When patient registration data is not present, then hide the next appointment card

* When patient summary screen is created, then load patient registration data

* Add `NextAppointment` feature flag

* When next appointment feature is enabled and patient registration data is present, then show next appointment card

* Update CHANGELOG