Scope `AppUpdateModule` to `TheActivityComponent`
In later commits, parts of `AppUpdateModule` will depend on `Country`, which is not possible to get before login. Since, we are currently using the dependencies from `AppUpdateModule` after login (in patients tab screen and settings screen), we can scope the update module to `TheActivity`. Our UI is also designed to show the update UI, after the app is logged in.