Use full text search when searching for patients (#3613)
* Stop rendering last seen date in patient search results

* Add `assignedFacilityName` to `PatientSearchResult` table

We will use this to replace last seen, will remove last seen in later commits.

* Add assigned facility ID and name as param in `PatientSearchResultViewModel`

* Remove assigned facility id from `SearchResult`

* Show assigned facility name when assigned facility is not same as current facility

* Remove last seen as param from `PatientSearchResultViewModel`

* Remove `LastSeen` from `PatientSearchResult`

* Redact `PatientSearchResult#toString`

* Ignore deleted patients when loading search results

* Remove `Patient` join statement when loading all patient in facility

Instead of joining the patient table, we are using the columns we already have in `PatientSearchResult` view

* Remove `Patient` join statement when loading search results by name

* Remove `Patient` join statement when loading search results by number

* Remove unused params from `PatientSearchResult`

We are only keeping the data we are consuming instead of adding unnecessary columns in the view. We can start adding to this data as required.

* Implement FTS tables for patient name, phone number and identifier

We require 3 different FTS tables, since we cannot create a FTS table for `PatientSearchResult` view, it has to be an entity. As such we have to use 3 different tables.

I have created an issue for having FTS support for views as well in Room.
https://issuetracker.google.com/issues/221895944

* Bump database version to 100

* Add query to search for patient by name or number

* Add migrations for `PatientSearchResult` view

This also contains migrations for FTS tables, `Patient`, `PatientPhoneNumber`, `BusinessID`

* Replace `PatientRepository#searchPagingSource` with `PatientRepository#search`

* Update CHANGELOG

* Clean up `PatientSearchResult#allPatientsInFacility`

* Rename `renderLastSeen` to `renderAssignedFacility` in `PatientSearchResultItemView`