Improve loading overdue patients query performance (#3625)
* Update `overdueInFacilityPagingSource` query

Instead of using the `OverdueAppointment` database view, we are building the query, so that we can filter the measurements and appointment based on facility to improve performance. We will do that in later commit.

* Filter measurements based on facility id when joining

We are only getting measurements from current facility or for patients that are assigned to current facility

* Remove `OverdueAppointment` as `DatabaseView`

* Add `assignedFacilityId` index for `Patient` table

* Bump database version to 102

* Add room migrations

* Update CHANGELOG