Only run jobs in `trigger_demo_release` workflow if it's a release branch
While the `demo_release` job script checks if the created branch is a release branch before triggering Bitrise build. Rest of the jobs don't do that. Since those are not that important jobs, a simple contains check is sufficient. We are going to continue check for the proper release branch in `TriggerBitriseRelease` script.