Remove marking older appointments for a patient as a side effect when scheduling an appointment (#3664)
* Add function to get all appointments for a patient

I know I can use `getAppointment` for a patient. But I wanted to check the entire list of appointments are marked as visited instead of one by one. This makes the tests a bit cleaner as well.

* Make `AppointmentRepository#markOlderAppointmentsAsVisited` as public

We are gonna move the side effect from the `schedule` function to the effect handler to make it explicit rather than hidden behaviour.

* When scheduling appointments from next button, mark older appointments as visited

* When scheduling appointments for a patient, then mark older appointments as visited

* Remove mark older appointments as visited side effect from `AppointmentRepository#schedule`

This is now done at effect handler level using `markOlderAppointmentsAsVisited` function

* Update CHANGELOG

* Update test names in `ScheduleAppointmentEffectHandlerTest`