Enable CDSS alerts only in pilot facilities (#3819)
* Provide CDSS pilot facilities list in `PatientSummaryModule`

This will allow us to fetch list of UUIDs from Firebase for facilities where the CDSS feature is enabled during pilot phase.

* Add effect to check if CDSS pilot is enabled for the facility

* When CDSS pilot is enabled for facility, then load clinical decision support info

* When CDSS pilot is not enabled for the facility, then do nothing

* When patient summary screen is created, then check if CDSS pilot is enabled

We will remove `isCdsAlertsFeatureEnabled` feature flag now, as the pilot facilities check will take care of enabling and disabling the feature for specific facilities.

* Remove `isCdsAlertsFeatureEnabled` as dependency for `PatientSummaryInit`

* Remove `isCdsAlertsFeatureEnabled` as dependency for `PatientSummaryViewRenderer`

* Remove `CDSAlerts` feature flag

* Update CHANGELOG

* Fix deeply nested control flow code climate issue

* Change remote config key for CDSS pilot facilities ids

Since this feature is not fully done. To be safe when enabling this feature, we will version the remote config key.