Migrate `LoggedOutOfDeviceDialog` to `BaseDialog` (#4174)
* Add screen key for `LoggedOutOfDeviceDialog`

* Make `BaseDialog#bindView` return value nullable

Not every dialog will have custom view attached to it. For that we don't need to have `bindView` as required value

* Extend `BaseDialog` in `LoggedOutOfDeviceDialog`

* Remove Mobius delegate from `LoggedOutOfDeviceDialog`

* Use screen key to open logged out of device dialog

* Remove unused code

* Use composite disposable for disposing events

Since there is a possibility the dialog is not handling events. Lateinit for events disposable never gets initialised. To avoid that we are adding the disposable we get when subscribing to events to the composite disposable

* Update CHANGELOG