Add color scheme for Tamil Nadu app (#4306)
* Add appbar layout theme overlay

* Remove blue0 primary theme overlay

We are primarily using this for theming app bar layout. Since we now have a widget specific theme overlay, we don't need this anymore

* Add manifest placeholder for app theme

This would allow us to change theme depending on product flavor we are building for. For example we can use `Theme.Simple` as default and for `tamilNadu` product flavor we can update the `appTheme` manifest placeholder to `Theme.Thanks`

* Use toolbar primary color for tab layout

* Update color scheme for Tamil Nadu app

* Ignore unused resources in lint

* Update CHANGELOG