Observe logged in user continuously in `UpdateAnalyticsUser`
This would enable us to set new analytics user whenever a user is logged out and logged in again.

But this would also mean every time user changes happen, this subscriber is triggered and Analytics user is updated. Since we are only consuming user ID in the AnalyticsUser, this should be fine.