Remove build status from the README (#4752)
This build status is for CI runs, but we don't run CI runs on push on
main branch, so the indicator is showing it for the last finished PR CI
run. It doesn't make sense to include that in the README. So, removing
it.

PS: Since it doesn't impact project in anyway, haven't update the
CHANGELOG