Stop clearing view binding reference when a custom view is detached
We initially did this because in documentation it was done for Fragments, but fragments outlive there views, so we were clearing it. Since that's not the case with views, it's not necessary.

In fact this can cause issues in scenarios where a view is detached because it's being recycled or in background, and once the view is attached again the binding is null and causes crashes