heroku setup: pass through seed exit status
We fix a bug in the setup script that always caused the exit code of the
rails db setup and seed to be successful, because it used the exit code
of `heroku run` itself. We add a flag that will instead pass through the
exit code of the underlying run, in this case the `rails` command.