heroku setup: pass through seed exit status (#4918)
We fix a bug in the setup script that always causes the exit code of the
rails db setup and seeding to be successful because it used the exit
code of `heroku run` itself. We add a flag that will instead pass
through the exit code of the underlying run, in this case, the `rails`
command.